common/libnetwork: prune dead network code
This code beacome unused with the migration to move most of this into
netavark in
commit fabe5060f3 ("Use netavark create to fill out network config")

Remove this so it does not confuse others who try to make changes to it.
I also inlined validateSetupOptions as this seems a bit simpler and we
actually have to only validate the network name there and not use the
loose matching logic as this would pass incorrect data to nv then in the
network array.
