+++ common/libnetwork/internal/util/validate.go
@@ -95,45 +95,3 @@ func ValidateSubnets(network *types.Network, addGateway bool, usedNetworks []*ne
 	}
 	return nil
 }
-
-func ValidateSetupOptions(n NetUtil, namespacePath string, options types.SetupOptions) error {
-	if namespacePath == "" {
-		return errors.New("namespacePath is empty")
-	}
-	if options.ContainerID == "" {
-		return errors.New("ContainerID is empty")
-	}
-	if len(options.Networks) == 0 {
-		return errors.New("must specify at least one network")
-	}
-	for _, net := range options.Networks {
-		network, err := n.Network(net.Name)
-		if err != nil {
-			return err
-		}
-		err = validatePerNetworkOpts(network, &net.PerNetworkOptions)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// validatePerNetworkOpts checks that all given static ips are in a subnet on this network.
-func validatePerNetworkOpts(network *types.Network, netOpts *types.PerNetworkOptions) error {
-	if netOpts.InterfaceName == "" {
-		return fmt.Errorf("interface name on network %s is empty", network.Name)
-	}
-	if network.IPAMOptions[types.Driver] == types.HostLocalIPAMDriver {
-	outer:
-		for _, ip := range netOpts.StaticIPs {
-			for _, s := range network.Subnets {
-				if s.Subnet.Contains(ip) {
-					continue outer
-				}
-			}
-			return fmt.Errorf("requested static ip %s not in any subnet on network %s", ip.String(), network.Name)
-		}
-	}
-	return nil
-}
